move revision diff files to dedicated revisionDiff dir
