+++ services/app-api/src/postgres/revisionDiff/index.ts
@@ -1,0 +1,6 @@
+export {
+    findRevisionDiffByContractID,
+    InvalidRevisionDiffInputError,
+    resolveRevisionPair,
+} from './findRevisionDiffByContractID'
+export type { FindRevisionDiffArgs } from './findRevisionDiffByContractID'
